create new page for screenshot
