+++ client/src/types/index.d.ts
@@ -1,6 +1,14 @@
 /// <reference types="vite/client" />
 
-type ConfigStep = "welcome" | "name" | "logo" | "color" | "completed" | "share" | "website-demo";
+type ConfigStep =
+  | "welcome"
+  | "name"
+  | "logo"
+  | "color"
+  | "completed"
+  | "share"
+  | "screenshot"
+  | "website-demo";
 
 type CardConfig = {
   name: string;
